feat: add new test show_rot_submenu function
